requirements.txt
feedparser==5.2.1
httplib2==0.9.2
python-slugify==1.2.0
Unidecode==0.4.19
